Optimize the parsing of large templates
We were getting performance issues for large templates with nested
custom tags and blocks.

In particular, items of the form:
```
{% block data %}
    {% simple data %}

    <large string>
{% endblock %}
```
Would parse a lot slower in liqp than shopify's ruby implementation,
roughly linearly with the size of the string addendum.

We've noticed that the SLL parser is blazing fast, but liqp falls back
to the LL parser a lot.

A case like
```
    {% block %}
      {% simple %}
    {% endblock %}
```
will use the LL parser because there isn't a matching end for `simple`,
whereas this works:
```
   {% block %}
      {% simple %}{% endsimple %}
    {% endblock %}
```

We augment the parser to work more dynamically -- with more information
about the starts/ends of tags and blocks, we can invoke the SLL parser, leading to
dramatic performance improvements.

Note that Liqp@HEAD does not differentiate between tags and blocks.

This PR significantly changes the API, since we now differentiate between
blocks and tags. However, since the tags and blocks are now passed to
the lexer for an optimized parser (SLL), we believe the changes are
worth it.

Also added some error handling for mismatched tags and broken templates.
Removed a few edges around open ({{) and close (}}) tags.

Closes #170.