Refactor Template parser to TemplateParser; fix global scope issue
Currently, there are several places in the API that may have unsafe
side-effects on otherwise unrelated code that shares Liqp running in the
same VM. This is due to global scope that allows modifications around
Insertions and Filters, which are stored as static HashMaps shared
across all callers.

In the worst case this could be seen as a security issue because
unrelated code may interfere and override a certain insertion (tag or
block), leading to information disclosure or data manipulation.

The bug may also just trigger a ConcurrentModificationException when
these modifications occur at an inopportune time.

This patch introduces a clean solution while providing backwards
compatibility for existing code. A new API class, TemplateParser, is
introduced, combining all settings-related configurations in one
immutable place.

Insertions and Filters are now maintained in immutable data structures,
and adding them via the global methods will not affect an existing
TemplateParser instance at all. In the previous cde, these were copied
multiple times from Lists to HashMaps during Template construction,
which is now also no longer necessary.

The Flavor enum is enhanced in such way that it can be used to quickly
access TemplateParsers/ParserSettings with default settings.

The old Template API is marked with @Deprecated statements, encouraging
users to adapt the new one.

Updated unit test code follows in a separate commit to verify that this
change works with the existing test code as well.