Allow keywords like "offset" in for/assign/capture/tablerow
Currently we require that the variable names in these blocks/tags are of
type "Id", which prevents use to use variable names like "offset". These
are commonly used in Jekyll and perfectly fine otherwise.

Change the parser to allow the existing type "id" (lower case) instead
of "Id" in these scenarios.