ruby: run_with*.sh should use realpath for shared folder
On macOS, the Docker app uses the canonical path (symlinks resolved) for
shared folders.

This prevents sharing the directory containing the ruby scripts if the
path of the current working directory contains any directory symlinks.

Use realpath instead of pwd in these scripts, falling back to using pwd
when realpath is not available.