Differentiate between null and missing values
Currently, we would get a VariableNotExistException when in strict mode,
even if a page property, for example, is knowingly set to null.

This is specifically important since {% if val %} returns true even for
empty strings, making it almost impossible to avoid an exception and
still support properties like page.last_modified_at.

Introduce a "found" state when traversing structures, and only consider
a non-existant variable if found=false

This change breaks backwards compatibility with classes implementing
LookupNode.Indexable. This is considered low-risk at the moment.